Moves toolbars back into viewer directory
