+++ src/lib/components/viewer/AnnotationToolbar.svelte
