+++ src/lib/components/viewer/stories/AnnotationToolbar.stories.svelte
