+++ src/lib/components/viewer/stories/RedactionToolbar.stories.svelte
